Implement basic blackjack strategy from CSV file
This commit adds a BasicStrategyLoader class that loads a basic blackjack strategy from a CSV file
and uses it to decide hit/stand actions.

The CSV file contains the standard basic strategy of when to hit, stand, double down, or split based
on the player's hand value and the dealer's up card. This data is loaded into a 2D array and indexed
to determine the action.

Fallback behavior is to stand if an invalid hand/upcard combination is requested. Default strategy
file location is checked into source control in a new 'files' directory. File location can also be
specified to allow flexible strategy definitions.